requirements.txt
# Communs aux deux scripts (doctosync.py et docto_heatmap.py)
PyYAML>=6.0.3
requests>=2.32.5
browser_cookie3>=0.20.1

# Spécifiques à doctosync.py (synchronisation Google Calendar)
google-api-python-client>=2.187.0
google-auth-oauthlib>=1.2.3

# Spécifiques à docto_heatmap.py (génération de heatmaps)
matplotlib>=3.9.0
seaborn>=0.13.0
pandas>=2.2.0

# Linting
pylint>=3.0.0
